feat: add stats server

+++ server/src/main/java/com/notbadcode/explorewithme/event/dto/EventShortDto.java
@@ -29,5 +29,5 @@ public class EventShortDto {
 
     Boolean paid;
 
-    int views;
+    Long views;
 }
